(#986) Refactor RunEngineSimulator so that tests work from the command line with pytest importlib
