+++ tests/unit_tests/device_setup_plans/test_dcm_pitch_roll_mirror_adjuster.py
@@ -11,19 +11,12 @@ from hyperion.device_setup_plans.dcm_pitch_roll_mirror_adjuster import (
     DCMPitchRollMirrorAdjuster,
 )
 
-from ..conftest import (
-    RunEngineSimulator,
-    assert_message_and_return_remaining,
-    mock_message_generator,
-)
-
 
 def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
-    dcm: DCM, vfm: VFM, hfm: HFM, qbpm1: QBPM1, s4_slit_gaps: S4SlitGaps, i0: I0
+    dcm: DCM, vfm: VFM, hfm: HFM, qbpm1: QBPM1, s4_slit_gaps: S4SlitGaps, i0: I0, sim
 ):
-    sim = RunEngineSimulator()
     peak_finder = MagicMock()
-    peak_finder.find_peak_plan = mock_message_generator("find_peak_plan")
+    peak_finder.find_peak_plan = sim.mock_message_generator("find_peak_plan")
     sim.add_handler(
         "read",
         "dcm_bragg_in_degrees",
@@ -36,7 +29,7 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
 
     messages = sim.simulate_plan(generator)
 
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages,
         lambda msg: msg.command == "find_peak_plan"
         and msg.args
@@ -48,10 +41,10 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
             0.002,
         ),
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:], lambda msg: msg.command == "sleep" and msg.args == (2,)
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "set"
         and msg.obj
@@ -59,7 +52,7 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
         and msg.args == (1.5,),
         "prepare_for_vfm_align",
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "set"
         and msg.obj
@@ -67,7 +60,7 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
         and msg.args == (0.017,),
         "prepare_for_vfm_align",
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "set"
         and msg.obj
@@ -75,7 +68,7 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
         and msg.args == (0.019,),
         "prepare_for_vfm_align",
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "set"
         and msg.obj
@@ -83,7 +76,7 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
         and msg.args == (500,),
         "prepare_for_vfm_align",
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "set"
         and msg.obj
@@ -91,10 +84,10 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
         and msg.args == (60,),
         "prepare_for_vfm_align",
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:], lambda msg: msg.command == "wait", "prepare_for_vfm_align"
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "find_peak_plan"
         and msg.args
@@ -106,7 +99,7 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
             0.0006,
         ),
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "set"
         and msg.obj
@@ -114,7 +107,7 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
         and msg.args == (100,),
         "prepare_for_hfm_align",
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "set"
         and msg.obj
@@ -122,10 +115,10 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
         and msg.args == (60,),
         "prepare_for_hfm_align",
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:], lambda msg: msg.command == "wait", "prepare_for_hfm_align"
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "find_peak_plan"
         and msg.args
@@ -137,7 +130,7 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
             0.0006,
         ),
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "set"
         and msg.obj
@@ -145,7 +138,7 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
         and msg.args == (500,),
         "reset_slit_gap",
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:],
         lambda msg: msg.command == "set"
         and msg.obj
@@ -153,6 +146,6 @@ def test_auto_adjust_dcm_hfm_vfm_pitch_roll(
         and msg.args == (500,),
         "reset_slit_gap",
     )
-    messages = assert_message_and_return_remaining(
+    messages = sim.assert_message_and_return_remaining(
         messages[1:], lambda msg: msg.command == "wait", "reset_slit_gap"
     )
